closes #8 add es_dangerous_reset --es-only flag

+++ django_elastic_migrations/models.py
@@ -3,21 +3,21 @@ from __future__ import (absolute_import, division, print_function, unicode_liter
 
 import datetime
 import json
+import os
+import random
 import sys
 import time
 import traceback
 from multiprocessing import cpu_count
 
-import os
-import random
 from django.db import models, transaction
 from django.utils import timezone
 from django.utils.encoding import python_2_unicode_compatible
 from elasticsearch import TransportError
 
-from django_elastic_migrations import codebase_id, environment_prefix, DEMIndexManager
+from django_elastic_migrations import codebase_id, environment_prefix, DEMIndexManager, es_client
 from django_elastic_migrations.exceptions import NoActiveIndexVersion, NoCreatedIndexVersion, IllegalDEMIndexState, \
-    CannotDropActiveVersion, IndexVersionRequired, CannotDropOlderIndexesWithoutForceArg
+    CannotDropActiveVersionWithoutForceArg, IndexVersionRequired, CannotDropOlderIndexesWithoutForceArg
 from django_elastic_migrations.utils.django_elastic_migrations_log import get_logger
 from django_elastic_migrations.utils.multiprocessing_utils import USE_ALL_WORKERS
 
@@ -80,7 +80,7 @@ class Index(models.Model):
         return self.get_available_versions().filter(prefix=prefix)
 
     def _get_other_versions(
-        self, given_version=None, prefix=environment_prefix, older=True):
+            self, given_version=None, prefix=environment_prefix, older=True):
         """
         Find all non-deleted versions that are older / newer
         than the given version. If older is False, return newer
@@ -154,7 +154,7 @@ class IndexVersion(models.Model):
 
     @property
     def is_deleted(self):
-        return not self.deleted_time
+        return bool(self.deleted_time)
 
     @property
     def name(self):
@@ -182,6 +182,21 @@ class IndexVersion(models.Model):
             parent_index.save()
         self.deleted_time = timezone.now()
         self.save()
+
+    def hard_delete(self):
+        self.delete()
+        super(IndexVersion, self).delete()
+
+    def get_schema_body(self):
+        body = json.loads(self.json)
+        return body
+
+    def get_indented_schema_body(self):
+        body = self.get_schema_body()
+        return json.dumps(body, sort_keys=True, indent=2)
+
+    def exists_in_es(self):
+        return es_client.indices.exists(index=self.name)
 
 
 @python_2_unicode_compatible
@@ -253,7 +268,10 @@ class IndexAction(models.Model):
         """
         Get a string representation of this model instance.
         """
-        return '<IndexAction, ID: {}>'.format(self.id)
+        index_name = self.index.name
+        if self.index_version:
+            index_name = self.index_version.name
+        return '<{} for {}, ID: {}>'.format(self.action, index_name, self.id)
 
     @property
     def dem_index(self):
@@ -573,7 +591,23 @@ class CreateIndexAction(IndexAction):
 
     def __init__(self, *args, **kwargs):
         self.force = kwargs.pop('force', False)
+        self.es_only = kwargs.pop('es_only', False)
         super(CreateIndexAction, self).__init__(*args, **kwargs)
+
+        # these task kwargs aren't used anywhere else,
+        # they are recorded in the database to make the history of actions
+        # taken on the indexes easier to understand
+        task_kwargs = {}
+        if self.force:
+            task_kwargs['force'] = self.force
+        if self.es_only:
+            task_kwargs['es_only'] = self.es_only
+
+        if task_kwargs:
+            self.task_kwargs = json.dumps(task_kwargs, sort_keys=True)
+            self.kwargs = task_kwargs
+        else:
+            self.kwargs = {}
 
     class Meta:
         # https://docs.djangoproject.com/en/2.0/topics/db/models/#proxy-models
@@ -585,7 +619,20 @@ class CreateIndexAction(IndexAction):
         self._index_name = self.index.name
 
         msg = ""
-        if latest_version and dem_index.hash_matches(latest_version.json_md5):
+        if self.es_only:
+            versions = self.index.get_available_versions_with_prefix()
+            for version in versions:
+                self._index_version_name = version.name
+                version_dem_index = DEMIndexManager.get_dem_index(self._index_version_name, exact_mode=True)
+                schema_body = version.get_schema_body()
+                created = version_dem_index.create_if_not_in_es(body=schema_body)
+                if created:
+                    self.add_log("{_index_version_name} wasn't found in Elasticsearch! We recreated it there with its original schema.", use_self_dict_format=True)
+                else:
+                    # TODO: check if schema matches and raise an exception if it does not
+                    self.add_log("{_index_version_name} was already created in Elasticsearch; did not create a new index.", use_self_dict_format=True)
+
+        elif latest_version and dem_index.hash_matches(latest_version.json_md5):
             self.index_version = latest_version
             self._index_version_name = self.index_version.name
             if self.force:
@@ -609,7 +656,9 @@ class CreateIndexAction(IndexAction):
                     self.add_log("   ... but it wasn't found in Elasticsearch! We recreated it there.")
                 else:
                     self.add_log("  ... did not create a new index.")
+
         else:
+            # either there is not an existing version, or it has changed, so just create a new one
             self.index_version = dem_index.create()
             self._index_version_name = self.index_version.name
             self.add_log(
@@ -689,7 +738,24 @@ class DropIndexAction(OlderModeMixin, IndexAction):
     def __init__(self, *args, **kwargs):
         self.force = kwargs.pop('force', False)
         self.just_prefix = kwargs.pop('just_prefix', None)
+        self.es_only = kwargs.pop('es_only', False)
+        self.hard_delete = kwargs.pop('hard_delete', False)
         super(DropIndexAction, self).__init__(*args, **kwargs)
+
+        # these task kwargs aren't used anywhere else, they are recorded in the database
+        # to make the history of actions taken on the indexes easier to understand
+        task_kwargs = {}
+        if self.force:
+            task_kwargs['force'] = self.force
+        if self.just_prefix:
+            task_kwargs['just_prefix'] = self.just_prefix
+        if self.es_only:
+            task_kwargs['es_only'] = self.es_only
+        if self.hard_delete:
+            task_kwargs['hard_delete'] = self.hard_delete
+
+        if task_kwargs:
+            self.task_kwargs = json.dumps(task_kwargs, sort_keys=True)
 
     class Meta:
         # https://docs.djangoproject.com/en/2.0/topics/db/models/#proxy-models
@@ -704,18 +770,26 @@ class DropIndexAction(OlderModeMixin, IndexAction):
                 versions = self.index.get_older_versions(
                     given_version=version_model, prefix=self.just_prefix)
                 action = DropIndexAction(
-                    force=self.force, just_prefix=self.just_prefix)
+                    force=self.force, just_prefix=self.just_prefix, es_only=self.es_only, hard_delete=self.hard_delete)
                 self.apply_to_older(versions, action=action)
             else:
                 self.index_version = version_model
-                if version_model.is_active:
-                    raise CannotDropActiveVersion()
+                if self.index_version.is_active and not self.force:
+                    raise CannotDropActiveVersionWithoutForceArg()
 
                 msg_params.update({"index_version_name": version_model.name})
-                dem_index.delete()
-                msg = ("Dropping index version '{index_version_name}' "
-                       "because you said to do so.".format(**msg_params))
-                self.add_log(msg)
+
+                if self.es_only:
+                    msg = "Dropping index version '{index_version_name}' ONLY in es because you said to do so.".format(**msg_params)
+                    self.add_log(msg)
+                    DEMIndexManager.delete_es_created_index(self.index_version.name, ignore=[404])
+                else:
+                    msg = "Dropping index version '{index_version_name}' because you said to do so.".format(**msg_params)
+                    self.add_log(msg)
+                    dem_index.delete()
+                    if self.hard_delete:
+                        self.index_version.hard_delete()
+                        self.index_version = None
         else:
             # first, check if *any* version exists.
             latest_version = self.index.get_latest_version()
@@ -749,34 +823,38 @@ class DropIndexAction(OlderModeMixin, IndexAction):
                 versions = self.index.get_older_versions(
                     given_version=self.index.active_version, prefix=self.just_prefix)
                 action = DropIndexAction(
-                    force=self.force, just_prefix=self.just_prefix)
+                    force=self.force, just_prefix=self.just_prefix, hard_delete=self.hard_delete)
                 return self.apply_to_older(versions, action=action)
+
             elif self.just_prefix:
                 available_versions = self.index.get_available_versions_with_prefix(self.just_prefix)
+
             else:
                 available_versions = self.index.get_available_versions()
-            self.add_log(
-                "About to drop {} versions because you said to do so "
-                "with the argument --force!".format(
-                    len(available_versions),
-                )
-            )
+
+            es_only_phrase = ""
+            if self.es_only:
+                es_only_phrase = "only in elasticsearch "
 
             # avoid circular import
-            from django_elastic_migrations import DEMIndexManager
             count = 0
 
             for version in available_versions:
                 count += 1
                 self.add_log(
-                    "Dropping version {} because you said to do so "
+                    "Dropping version {name} {es_only_phrase}because you said to do so "
                     "with the argument --force!".format(
-                        version.name
+                        name=version.name, es_only_phrase=es_only_phrase
                     )
                 )
                 try:
                     DEMIndexManager.delete_es_created_index(version.name)
-                    version.delete()
+                    if not self.es_only:
+                        if self.hard_delete:
+                            version.hard_delete()
+                            self.index_version = None
+                        else:
+                            version.delete()
                 except TransportError as ex:
                     if ex.status_code == 404:
                         self.add_log(
@@ -784,20 +862,24 @@ class DropIndexAction(OlderModeMixin, IndexAction):
                                 version.name
                             )
                         )
-                        if version.is_deleted:
-                            self.add_log(
-                                "Version {} was already deleted in IndexVersion model; "
-                                "not doing anything further".format(
-                                    version.name
+                        if not self.es_only:
+                            if self.hard_delete:
+                                version.hard_delete()
+                                self.index_version = None
+                            elif version.is_deleted:
+                                self.add_log(
+                                    "Version {} was already deleted in IndexVersion model; "
+                                    "not doing anything further".format(
+                                        version.name
+                                    )
                                 )
-                            )
-                            count -= 1
-                        else:
-                            version.delete()
+                                count -= 1
+                            else:
+                                version.delete()
                     else:
                         raise ex
 
-            self.add_log("Done dropping {} versions.".format(count))
+            self.add_log("Done dropping {count} versions {es_only_phrase}".format(count=count, es_only_phrase=es_only_phrase))
 
 
 class UpdateIndexAction(NewerModeMixin, IndexAction):
@@ -1067,7 +1149,7 @@ class PartialUpdateIndexAction(UpdateIndexAction):
             # if we're less than 10% done, put a little randomness
             # in between the workers to dither query load
             if (self.parent.docs_affected / self._total_docs_expected) < 0.1:
-                time.sleep(random.random()*2)
+                time.sleep(random.random() * 2)
 
         qs = doc_type.get_queryset()
         current_qs = qs[start:end]
@@ -1121,7 +1203,7 @@ class PartialUpdateIndexAction(UpdateIndexAction):
         self._parent_docs_affected_pct = 100 - self._total_docs_remaining_pct
         if self._parent_docs_per_sec:
             self._expected_parent_runtime = str(datetime.timedelta(
-                seconds= self._total_docs_remaining / self._parent_docs_per_sec))
+                seconds=self._total_docs_remaining / self._parent_docs_per_sec))
         else:
             self._expected_parent_runtime = 'unknown'
